Register + Verify handlers

+++ Common/Database/Models/Account.cs
@@ -6,14 +6,14 @@ namespace Common.Database.Models;
 public class Account {
     public int Id { get; set; }
     public string Name { get; set; }
-    public short Rank { get; set; }
+    public int Rank { get; set; }
     public string GuildName { get; set; }
     public bool IsAdmin { get; set; }
     public bool IsBanned { get; set; }
     public bool IsMuted { get; set; }
-    public short MaxChars { get; set; }
-    public short VaultCount { get; set; }
-    public short NextCharId { get; set; }
+    public int MaxChars { get; set; }
+    public int VaultCount { get; set; }
+    public int NextCharId { get; set; }
     public DateTime CreatedAt { get; set; }
     public AccountStats Stats { get; set; }
     public AccountGifts Gifts { get; set; }
